Unify translation utilities
